:recycle: [UploadableTrait] Add trait
